Extract new interfaces, classes for chains into separate files
