+++ src/commonMain/kotlin/ehn/techiop/hcert/kotlin/chain/IChain.kt
@@ -1,0 +1,37 @@
+package ehn.techiop.hcert.kotlin.chain
+
+import ehn.techiop.hcert.kotlin.data.GreenCertificate
+import kotlin.js.JsName
+
+/**
+ * Interface for the encoding / decoding chain
+ */
+interface IChain {
+    /**
+     * Process the [input], apply encoding in this order:
+     * - [CborService]
+     * - [CwtService]
+     * - [CoseService]
+     * - [CompressorService]
+     * - [Base45Service]
+     * - [ContextIdentifierService]
+     *
+     * The result ([ChainResult]) will contain all intermediate steps, as well as the final result in [ChainResult.step5Prefixed].
+     */
+    @JsName("encode")
+    fun encode(input: GreenCertificate): ChainResult
+
+    /**
+     * Process the [input], apply decoding in this order:
+     * - [ContextIdentifierService]
+     * - [Base45Service]
+     * - [CompressorService]
+     * - [CoseService]
+     * - [CwtService]
+     * - [CborService]
+     * - [SchemaValidationService]
+     * The result ([ChainDecodeResult]) will contain the parsed data, as well as intermediate results.
+     */
+    @JsName("decode")
+    fun decode(input: String): DecodeResult
+}
